Fixed Emscripten compiler errors.

+++ src/SDG/Game/Datatypes/Tiled/ChunkSize.cpp
@@ -2,7 +2,7 @@
 
 void SDG::Tiled::ChunkSize::LoadImpl(const Xml::XmlElement &el)
 {
-    size_t width = 16, height = 16;
+    unsigned width = 16, height = 16;
     el.Attribute("width", Xml::Required).Query(width);
     el.Attribute("height", Xml::Required).Query(height);
